containers.hashmap: Add getOrAdd method
getOrAdd returns a pointer to the value inside the HashMap if one
exists for the given key. If not, it inserts a new key/value
pair (using the given key and defaultValue), and returns a pointer to
the added value.

This method is necessary to efficiently implement operations such as
the builtin AAs' `aa[key]++` when `key` is not yet present in `aa`.