Array.insert: don't dtor or opAssign uninitialized struct
For structs, if a destructor is run before blit or copying involves
more than just a blit, ensure that arr[l] is in a valid state before
assigning to it.

This patch performs the initialization lazily right before insertion.
Another approach would be to instead initialize the memory when the
array is allocated/expanded by using std.experimental.allocator:
makeArray/expandArray.