containers.hashmap: Re-add the storeHash parameter
In some cases, the hash can be extremely cheap to calculate (e.g. the
first few bytes of the key, when the key itself is a longer hash). In
these cases, storing the hash is an unnecessary overhead. As such, the
user should be able to specify when to store the hash within HashMap
nodes.

The storeHash parameter was initially introduced in
fb998f0c440529f78d54f534e7407741355d46f7, and was later
removed (presumably accidentally) in
15eea7fbc89f073838d14d86d47c1247dbe8c3a1. As it was part of the API,
this was a breaking regression. This commit Fixes the regression and
adds the parameter again.