containers.hashmap: Fix redundant hashFunction call in getOrAdd
The result was being thrown away by being overwritten inside
find(). Additionally, when a hash function was specified with a return
value other than size_t, this resulted in a compilation error.

The bug is a regression (introduced in commit
15eea7fbc89f073838d14d86d47c1247dbe8c3a1).

Fix by letting find compute the hash as in other invocations.