Fix writing to dangling pointers to stack
Unittest helper types S, C, Cls hold a pointer to a counter which
tracks the number of destructions. However, in some circumstances,
these destructions may be non-deterministic, i.e. occur as a result of
a garbage collection cycle. Should this occur after the unittest has
already exited, then the pointer held by these types will be invalid.

This causes a cyclicbuffer unittest to segfault when compiled with
LDC.

Fix this by always allocating the counter in the managed heap, so that
its lifetime is properly tracked by the garbage collector, along with
the test variables.