CI fix rollup (#180)
* fix: use std.experimental allocator

This removes the dependency on `stdx.allocator` by using the built-in `std.experimental.allocator`.

* containers.unrolledlist: Fix unittest on LDC

Patch based on https://github.com/dlang-community/containers/issues/164#issue-823151444

Fixes #164.

* Fix writing to dangling pointers to stack

Unittest helper types S, C, Cls hold a pointer to a counter which
tracks the number of destructions. However, in some circumstances,
these destructions may be non-deterministic, i.e. occur as a result of
a garbage collection cycle. Should this occur after the unittest has
already exited, then the pointer held by these types will be invalid.

This causes a cyclicbuffer unittest to segfault when compiled with
LDC.

Fix this by always allocating the counter in the managed heap, so that
its lifetime is properly tracked by the garbage collector, along with
the test variables.

Co-authored-by: Amin Yahyaabadi <aminyahyaabadi74@gmail.com>